LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/inline/set-shorthand-expected.txt


FAIL Setting a shorthand with an invalid CSSStyleValue on inline style throws TypeError assert_throws_js: function "() => styleMap.set(property, ...values)" threw object "NotSupportedError: set() is not yet supported" ("NotSupportedError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Setting a shorthand with an invalid String on inline style throws TypeError assert_throws_js: function "() => styleMap.set(property, ...values)" threw object "NotSupportedError: set() is not yet supported" ("NotSupportedError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Setting a shorthand with a CSSStyleValue updates inline style set() is not yet supported
FAIL Setting a shorthand with a string updates inline style set() is not yet supported
